+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 ### IMPROVEMENTS
 
+- `[consensus]` reuse encode/decode buffers in WALEncoder and WALDecoder.
+  ([\#5865](https://github.com/cometbft/cometbft/pull/5865))
 - `[blocksync]` validate blocksync response sender and signature count
   ([\#5860](https://github.com/cometbft/cometbft/pull/5860))